// front/src/types/settings.ts
export interface SiteSettings {
  showMobileLogo: number;
  visitorTracking: number;
  currency: string;
  logoUrl: string;

  cardSize: number;
  imageSize: number;

  mobileCardSize: number;
  mobileImageSize: number;

  tabCardSize: number;
  tabImageSize: number;

  cardBg: string;
  imageBg: string;
  textColor: string;

  showTopBanner: number;
  bannerUrl: string;
  bannerAnimation: string;

  bgColor: string;
  bgShadowColor: string;

  mobFontColor: string;
  lapFontColor: string;

  mobFtWeight: number;
  lapFtWeight: number;

  mobFtSize: number;
  lapFtSize: number;

  mobHeight: string;
  lapHeight: string;

  autoSlideHero: number;
  primaryColor: string;
  secondaryColor: string;
  thirdColor: string;

  button1Color: string;
  button2Color: string;
  button3Color: string;
  
  button1TextColor: string;
  button2TextColor: string;
  button3TextColor: string; 
  
  starColor1: string;
  starColor2: string;
}